chore: rename ticket service file
