+++ services/api-main/src/app/tickets/domain/services/ticket.service.ts
